+++ src/Random/RandomByteGeneratorInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Drenso\GenPhrase\Random;
+
+interface RandomByteGeneratorInterface
+{
+  public function getBytes(int $count): string;
+}